requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
